+++ tests/ValidatorTest.php
@@ -22,13 +22,13 @@ use Yiisoft\Validator\Rule\Required;
 use Yiisoft\Validator\RuleInterface;
 use Yiisoft\Validator\SimpleRuleHandlerContainer;
 use Yiisoft\Validator\SkipOnEmptyCallback\SkipOnNull;
-use Yiisoft\Validator\Tests\Stub\DataSet;
+use Yiisoft\Validator\Tests\Support\DataSet\SimpleDataSet;
 use Yiisoft\Validator\Tests\Support\ValidatorFactory;
 use Yiisoft\Validator\Tests\Support\Rule\NotNullRule\NotNull;
 use Yiisoft\Validator\Tests\Support\Data\ObjectWithDataSet;
 use Yiisoft\Validator\Tests\Support\Rule\StubRule\StubRule;
 use Yiisoft\Validator\Tests\Support\Data\ObjectWithAttributesOnly;
-use Yiisoft\Validator\Tests\Stub\TranslatorFactory;
+use Yiisoft\Validator\Tests\Support\TranslatorFactory;
 use Yiisoft\Validator\ValidationContext;
 use Yiisoft\Validator\Validator;
 
@@ -36,7 +36,7 @@ class ValidatorTest extends TestCase
 {
     public function testAddingRulesViaConstructor(): void
     {
-        $dataObject = new DataSet(['bool' => true, 'int' => 41]);
+        $dataObject = new SimpleDataSet(['bool' => true, 'int' => 41]);
         $validator = ValidatorFactory::make();
         $result = $validator->validate($dataObject, [
             'bool' => [new Boolean()],
@@ -96,7 +96,7 @@ class ValidatorTest extends TestCase
     {
         $validator = ValidatorFactory::make();
         $result = $validator->validate(
-            new DataSet(['property' => '']),
+            new SimpleDataSet(['property' => '']),
             ['property' => [new Required(when: static fn (mixed $value, ?ValidationContext $context): bool => false)]],
         );
 
@@ -110,7 +110,7 @@ class ValidatorTest extends TestCase
         $validator = ValidatorFactory::make();
 
         $this->expectException(RuleHandlerInterfaceNotImplementedException::class);
-        $validator->validate(new DataSet(['property' => '']), [
+        $validator->validate(new SimpleDataSet(['property' => '']), [
             'property' => [
                 new class ($ruleHandler) implements RuleInterface {
                     public function __construct(private $ruleHandler)
@@ -136,7 +136,7 @@ class ValidatorTest extends TestCase
         $this->expectException(RuleHandlerNotFoundException::class);
 
         $validator = ValidatorFactory::make();
-        $validator->validate(new DataSet(['property' => '']), [
+        $validator->validate(new SimpleDataSet(['property' => '']), [
             'property' => [
                 new class () implements RuleInterface {
                     public function getName(): string
